Fix clang compilation errors (#119)
**Public-Facing Changes**
None


**Description**
- Fix clang compilation errors
  - [x] ROS2
  - [x] ROS1

Fixes #118

+++ ros2_foxglove_bridge/src/ros2_foxglove_bridge.cpp
@@ -153,7 +153,7 @@ public:
       this->set_parameter(rclcpp::Parameter{"port", listeningPort});
     }
 
-    _maxQosDepth = this->get_parameter("max_qos_depth").as_int();
+    _maxQosDepth = static_cast<size_t>(this->get_parameter("max_qos_depth").as_int());
 
     // Start the thread polling for rosgraph changes
     _rosgraphPollThread =
